feat: refactor borrowers avatars component for thank you page single version variant
